updated permissions

+++ src/controllers/movies/index.js
@@ -22,7 +22,7 @@ const movieController = {
         }
     },
 
-    // 1) addMovie
+    // 2) addMovie
     addMovie: async (req, res) => {
         
         try {
@@ -51,14 +51,24 @@ const movieController = {
         }
     },
 
+    // 3) editMovieByIdPage
     editMovieByIdPage :  async (req, res)=>{
 
         try {
+
+            // req.app_user
             
             const movie = await pool.query(
-                "SELECT * FROM movie WHERE id = $1",
-                [ req.params._id  ]
+                "SELECT * FROM movie WHERE id = $1 AND app_user_id = $2",
+                [ req.params._id , req.app_user.id ]
             )
+
+            console.log( "movie" , movie )
+
+            if( movie.rows.length === 0 ){
+                req.flash('error_msg', "Access denied!!")
+                return res.redirect("/")
+            }
 
             movie.rows[0].movie_cast =  movie.rows[0].movie_cast?.reduce((curr, cast)=>{
                   return curr + cast + ","
@@ -80,7 +90,7 @@ const movieController = {
     },
 
 
-    // 2) editMovieById
+    // 4) editMovieById
     editMovieById: async (req, res) => {
     
         try {
@@ -96,9 +106,15 @@ const movieController = {
             const casts = req.body.movie_cast.split(",")
 
             const movie = await pool.query(
-                "UPDATE movie SET name = $1 , rating = $2 , movie_cast = $3 , genre = $4 , release_date = $5 WHERE id = $6 RETURNING *",
-                [ name , rating , casts , genre , release_date , req.params._id  ]
+                "UPDATE movie SET name = $1 , rating = $2 , movie_cast = $3 , genre = $4 , release_date = $5 WHERE id = $6 AND app_user_id = $7 RETURNING *",
+                [ name , rating , casts , genre , release_date , req.params._id , req.app_user.id ]
             )
+
+            if( movie.rows.length === 0 ){
+                req.flash('error_msg', "Access denied!!")
+                return res.redirect("/")
+            }
+
             req.flash('success_msg', 'Movie Successfully Edited!')
             res.redirect("/")
         } catch (error) {
@@ -115,10 +131,18 @@ const movieController = {
 
             console.log(  "delete movie" )
 
-            const movies = await pool.query(
-                "DELETE FROM movie WHERE id = $1",
-                [ req.params._id ]
+            const movie = await pool.query(
+                "DELETE FROM movie WHERE id = $1 AND app_user_id = $2",
+                [ req.params._id , req.app_user.id ]
             )
+
+            console.log( movie.rows )
+
+            if( movie.rows.length === 0 ){
+                req.flash('error_msg', "Access denied!!")
+                return res.redirect("/")
+            }
+
             req.flash('success_msg', 'Movie Successfully Deleted!')
             res.redirect("/")
         } catch (error) {
